Add query price and place order logic

+++ app/src/main/java/com/mining/mining/mining/modules/home/model/bean/MiningDifficulty.kt
@@ -3,5 +3,5 @@ package com.mining.mining.mining.modules.home.model.bean
 data class MiningDifficulty(
         var difficulty: String = "",
         var prediction: String = "",
-        var update_time: Int = 0
+        var update_time: Long = 0
 )
